增加对回复异常的判断，完善回复Commentcount增加

+++ src/main/java/com/lchcommunity/community/controller/PublishController.java
@@ -28,7 +28,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String publish(@PathVariable("id") Integer id,
+    public String publish(@PathVariable("id") Long id,
                           Model model) {
         QuestionDTO question = questionService.getQuestionById(id);
         model.addAttribute("title", question.getTitle());
@@ -43,7 +43,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model
     ) {
